requirements.txt
altgraph==0.17.2
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
cryptography==37.0.2
cycler==0.11.0
fonttools==4.33.3
google-images-download==2.8.0
h11==0.13.0
idna==3.3
kiwisolver==1.4.2
macholib==1.16
numpy==1.22.3
outcome==1.1.0
packaging==21.3
Pillow==9.1.0
pycparser==2.21
pyinstaller-hooks-contrib==2022.4
pyOpenSSL==22.0.0
pyparsing==3.0.8
PySocks==1.7.1
python-dateutil==2.8.2
selenium==4.1.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
wsproto==1.1.0
